requirements/dev.txt
-r base.txt

-e .

pytest-django==2.6.2
pytest==2.6.2
py==1.4.24
pytest-cov==1.8.0
cov-core==1.14.0
coverage==3.7.1

factory_boy==2.4.1

WebOb==1.4
WebTest==2.0.16
beautifulsoup4==4.3.2
django-webtest==1.7.7
waitress==0.8.9
